Attempt exercise 3 for flex section
Made the index page follow specifications of the readme instructions.
Page looks like the outcome images and shares behavior to the provided
gif. Also, set flex-shrink to 0 for both the notification button and
profile pic even though not specified since it appears to maintain its
shape its shape in the gif.